# Speed up HPXML files w/ multiple `Building` elements