--- a/HPXMLtoOpenStudio/measure.xml
+++ b/HPXMLtoOpenStudio/measure.xml
@@ -3,8 +3,8 @@
   <schema_version>3.0</schema_version>
   <name>hpxm_lto_openstudio</name>
   <uid>b1543b30-9465-45ff-ba04-1d1f85e763bc</uid>
-  <version_id>7e74ed1d-aafd-41e1-9614-76cda6d48803</version_id>
-  <version_modified>20230323T222109Z</version_modified>
+  <version_id>7b6192c5-a5a0-4ba7-a3d0-994c401a4dc6</version_id>
+  <version_modified>20230328T223912Z</version_modified>
   <xml_checksum>D8922A73</xml_checksum>
   <class_name>HPXMLtoOpenStudio</class_name>
   <display_name>HPXML to OpenStudio Translator</display_name>
@@ -541,12 +541,6 @@
       <checksum>3FD8D019</checksum>
     </file>
     <file>
-      <filename>hpxml.rb</filename>
-      <filetype>rb</filetype>
-      <usage_type>resource</usage_type>
-      <checksum>A40418CD</checksum>
-    </file>
-    <file>
       <filename>test_schedules.rb</filename>
       <filetype>rb</filetype>
       <usage_type>test</usage_type>
@@ -581,6 +575,12 @@
       <filetype>csv</filetype>
       <usage_type>resource</usage_type>
       <checksum>1362DD15</checksum>
+    </file>
+    <file>
+      <filename>hpxml.rb</filename>
+      <filetype>rb</filetype>
+      <usage_type>resource</usage_type>
+      <checksum>CE61207A</checksum>
     </file>
   </files>
 </measure>