--- a/workflow/sample_files/base-multiple-buildings.xml
+++ b/workflow/sample_files/base-multiple-buildings.xml
@@ -447,14 +447,6 @@
           <LabelUsage>6.0</LabelUsage>
           <Capacity>3.2</Capacity>
         </ClothesWasher>
-        <ClothesDryer>
-          <SystemIdentifier id='ClothesDryer1'/>
-          <Location>living space</Location>
-          <FuelType>electricity</FuelType>
-          <CombinedEnergyFactor>3.73</CombinedEnergyFactor>
-          <Vented>true</Vented>
-          <VentedFlowRate>150.0</VentedFlowRate>
-        </ClothesDryer>
         <Dishwasher>
           <SystemIdentifier id='Dishwasher1'/>
           <Location>living space</Location>
@@ -982,14 +974,6 @@
           <LabelUsage>6.0</LabelUsage>
           <Capacity>3.2</Capacity>
         </ClothesWasher>
-        <ClothesDryer>
-          <SystemIdentifier id='ClothesDryer1_2'/>
-          <Location>living space</Location>
-          <FuelType>electricity</FuelType>
-          <CombinedEnergyFactor>3.73</CombinedEnergyFactor>
-          <Vented>true</Vented>
-          <VentedFlowRate>150.0</VentedFlowRate>
-        </ClothesDryer>
         <Dishwasher>
           <SystemIdentifier id='Dishwasher1_2'/>
           <Location>living space</Location>
@@ -1517,14 +1501,6 @@
           <LabelUsage>6.0</LabelUsage>
           <Capacity>3.2</Capacity>
         </ClothesWasher>
-        <ClothesDryer>
-          <SystemIdentifier id='ClothesDryer1_3'/>
-          <Location>living space</Location>
-          <FuelType>electricity</FuelType>
-          <CombinedEnergyFactor>3.73</CombinedEnergyFactor>
-          <Vented>true</Vented>
-          <VentedFlowRate>150.0</VentedFlowRate>
-        </ClothesDryer>
         <Dishwasher>
           <SystemIdentifier id='Dishwasher1_3'/>
           <Location>living space</Location>