## Pull Request Description

Speed up schematron validation when running simulations for HPXML files with multiple `Building` elements; now we only do schematron validation on the single `Building` element of interest. This also means we now get warnings specific to the single `Building` of interest too, which is desirable.

For a test HPXML file with 21 `Building` elements:

  | Schematron Validation runtime | Total runtime
-- | -- | --
Before | 2.2s | 7.9s
After | 0.3s | 6.0s

## Checklist

PR Author: Check these when they're done. Not all may apply. ~~strikethrough~~ and check any that do not apply. 

PR Reviewer: Verify each has been completed.

- [x] ~Schematron validator (`EPvalidator.xml`) has been updated~
- [x] ~Sample files have been added/updated (via `tasks.rb`)~
- [x] Tests have been added/updated (e.g., `HPXMLtoOpenStudio/tests` and/or `workflow/tests/hpxml_translator_test.rb`)
- [x] ~Documentation has been updated~
- [x] Changelog has been updated
- [x] `openstudio tasks.rb update_measures` has been run
- [x] No unexpected changes to simulation results of sample files